+++ openmetadata-ui/src/main/resources/ui/src/axiosAPIs/miscAPI.ts
@@ -18,7 +18,6 @@ import { WILD_CARD_CHAR } from '../constants/char.constants';
 import { SearchIndex } from '../enums/search.enum';
 import { AirflowConfiguration } from '../generated/configuration/airflowConfiguration';
 import { AuthenticationConfiguration } from '../generated/configuration/authenticationConfiguration';
-import { ResourcePermission } from '../generated/entity/policies/accessControl/resourcePermission';
 import { EntitiesCount } from '../generated/entity/utils/entitiesCount';
 import { Paging } from '../generated/type/paging';
 import { getURLWithQueryFields } from '../utils/APIUtils';
@@ -119,18 +118,6 @@ export const deleteLineageEdge: Function = (
   return APIClient.delete(
     `/lineage/${fromEntity}/${fromId}/${toEntity}/${toId}`
   );
-};
-
-export const getLoggedInUserPermissions = async () => {
-  const params = {
-    limit: 100,
-  };
-  const response = await APIClient.get<{
-    data: ResourcePermission[];
-    paging: Paging;
-  }>('/permissions', { params });
-
-  return response.data;
 };
 
 export const getInitialEntity = (
